Better dots placement in some cases
If one node is handle and the other is smooth node, then the next node on line segment will be used instead of smooth node. In this case only the dot around handle will be drawn an it will be located along the straight line. It improves better placement of directional dots.